+++ bin/install-chromedriver.sh
@@ -1,7 +1,5 @@
 #!/usr/bin/env bash
 
-# http://chromedriver.chromium.org/
-
 set -e
 
 if [ $# != 0 ]; then
@@ -12,9 +10,20 @@ fi
 # based on info in http://chromedriver.chromium.org/downloads
 # choose the chromedriver version based on the chrome version
 
-GOOGLE_CHROME_VERSION=$(google-chrome --version)
-echo "Version >>>${GOOGLE_CHROME_VERSION}<<< of Google Chrome is installed"
-case $(echo "${GOOGLE_CHROME_VERSION}" | sed -e 's|^Google Chrome ||g' | sed -e 's|\..*$||g') in
+if [[ "$(which google-chrome || true)" != "" ]]; then
+    # "google-chrome --version" returns something like "Google Chrome 87.0.4280.141"
+    BROWSER_VERSION=$(google-chrome --version)
+    BROWSER_MAJOR_VERSION_NUMBER=$(echo "${BROWSER_VERSION}" | sed -e 's|^Google Chrome ||g' | sed -e 's|\..*$||g')
+elif [[ "$(which chromium-browser)" != "" ]]; then
+    # "chromium-browser --version" returns something like "Chromium 87.0.4280.66 Built on Ubuntu , running on Ubuntu 18.04"
+    BROWSER_VERSION=$(chromium-browser --version)
+    BROWSER_MAJOR_VERSION_NUMBER=$(echo "${BROWSER_VERSION}" | sed -e 's|^[^0-9]*||g' | sed -e 's|\..*$||g')
+else
+    echo "Could not find chrome or chromium so chromedriver not installed." >&2
+    exit 1
+fi
+
+case "${BROWSER_MAJOR_VERSION_NUMBER}" in
 
   85)
     CHROMEDRIVER_VERSION=85.0.4183.87
@@ -29,7 +38,7 @@ case $(echo "${GOOGLE_CHROME_VERSION}" | sed -e 's|^Google Chrome ||g' | sed -e 
     ;;
 
   *)
-    echo "Don't know what to do with version >>>${GOOGLE_CHROME_VERSION}<<< of Google Chrome" >&2
+    echo "Don't know what to do with browser version >>>${BROWSER_VERSION}<<<" >&2
     exit 1
     ;;
 esac