+++ CHANGELOG.md
@@ -15,6 +15,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * ```cloudfeaster-lite``` docker image is now based on Alpine 3.12 (used to be Alpine 3.8)
 * ```install-chrome.sh``` now able to install both [Chrome](https://www.google.com/chrome/)
   and [Chromium](https://www.chromium.org/Home) based on command line switches
+* ```install-chromedriver.sh``` determines which version of chromedriver to install based
+  on which version of [Chrome](https://www.google.com/chrome/) or [Chromium](https://www.chromium.org/Home)
+  is installed
 
 ### Removed